% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{computeCriterion}
\alias{computeCriterion}
\title{Compute Internal Clustering Validity Criterion}
\usage{
computeCriterion(data, clustering, criterion)
}
\arguments{
\item{data}{Numeric matrix of observations (e.g., term trajectories).}

\item{clustering}{Integer vector of cluster assignments.}

\item{criterion}{Character. Name of the criterion to compute.}
}
\value{
A numeric value corresponding to the criterion.
}
\description{
Internal utility function to compute a specified internal clustering validation index,
using either the \pkg{clusterCrit} or \pkg{clusterSim} packages.
}
