% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRandMatrix.R
\name{computeRandMatrix}
\alias{computeRandMatrix}
\title{Compute Adjusted Rand Index Matrix for Selected Partitions}
\usage{
computeRandMatrix(best_partitions)
}
\arguments{
\item{best_partitions}{An object of class \code{ccccBestPartitions}, as returned by \code{selectBestPartitions()}.}
}
\value{
A list with the following elements:
\describe{
\item{rand_matrix}{A symmetric matrix of pairwise Adjusted Rand Index values.}
\item{partition_labels}{A character vector of partition names used as row/column names.}
\item{partitions}{A named list of all cluster partitions.}
\item{partition_k}{A named vector indicating the number of clusters (\code{k}) for each partition.}
}
}
\description{
This function computes the Adjusted Rand Index (ARI) between all pairs of cluster partitions
selected in the \code{ccccBestPartitions} object. The ARI measures the similarity between two
partitions, correcting for chance.
}
\examples{
\dontrun{
tdm <- system.file("extdata", "tdm.csv", package = "cccc")
corpus <- system.file("extdata", "corpus.csv", package = "cccc")
data <- importData(tdm_file = tdm, corpus_file = corpus,
sep_tdm = ";",sep_corpus_info = ";",zone="stat")

data_nchi <- normalization(data, normty = "nchi", sc = 1000)
g <- runClusteringRange(data_nchi,
k_range = 2:26,
n_repeats = 20,
seed = 123,
dist_method = "euclidean",
verbose = TRUE)

g2 <- buildCIVf(clustering_set=g,
criteria_set = NULL,
min_valid_values = 0.75,
verbose = TRUE)

g3 <- selectBestPartitions(g2, g, graph=T)
rand_result <- computeRandMatrix(g3)
}

}
