% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOptimalFits.R
\name{plotOptimalFits}
\alias{plotOptimalFits}
\title{Plot Optimal Smoothed Curves for Selected Keywords}
\usage{
plotOptimalFits(
  data,
  smoothed = NULL,
  opt_res = NULL,
  n_curves = 9,
  show_zone = FALSE,
  graph = FALSE
)
}
\arguments{
\item{data}{A list containing the outputs returned by \code{\link[=importData]{importData()}}}

\item{smoothed}{The functional data object, which is the smoothed spline fit based on the optimal parameters, returned by \code{\link[=plotOptimalSmoothing]{plotOptimalSmoothing()}}. If it is \code{NULL} argument \code{opt_res} must be provided.}

\item{opt_res}{A list with the optimal smoothing configuration returned by \code{\link[=optimalSmoothing]{optimalSmoothing()}}. If it is \code{NULL} argument \code{smoothed} must be provided.}

\item{n_curves}{Integer. Number of keywords to display, chosen across the RMS distribution (default is 9).}

\item{show_zone}{Logical. If \code{TRUE}, include the frequency zone (if available in \code{tdm}) in the plot labels.}

\item{graph}{Logical. If \code{FALSE} (default), print the sole combined plot to the active device. If \code{TRUE}, print all individual plots as well.}
}
\value{
A named list with the following elements (returned invisibly):
\itemize{
\item \code{singleKeywordPlot}: A list of ggplot objects, one for each selected keyword (full title included).
\item \code{combinedKeywordPlot}: A combined \code{patchwork} plot showing all selected fits.
}
}
\description{
This function visualizes raw and smoothed time trends of keyword frequencies based on smoothing spline fits.
A subset of keywords is selected based on the residual Root Mean Square (RMS) error of the smoothed fit.
Both individual and combined plots are returned.
}
\details{
The function requires the smoothed spline fit based on the optimal parameters, which is either provided directly via the \code{smoothed} argument or computed internally using the \code{opt_res} argument.
It then computes the residual RMS for each keyword and selects a subset using quantiles.
Each keyword's frequency curve (raw and smoothed) is plotted, either individually or in a combined layout using the \code{patchwork} package.
The individual plots include detailed titles with keyword, RMS value, RMS-based position, and optionally the frequency zone.
}
\examples{
\dontrun{
tdm <- system.file("extdata", "tdm.csv", package = "cccc")
corpus <- system.file("extdata", "corpus.csv", package = "cccc")
data <- importData(tdm_file = tdm, corpus_file = corpus,
sep_tdm = ";",sep_corpus_info = ";",zone="stat")

data_nchi <- normalization(data, normty = "nchi", sc = 1)

res_list <- list(
  "m-2" = smoothingSelection(data_nchi, penalty_type = "m-2", plot = FALSE),
  "2"   = smoothingSelection(data_nchi, penalty_type = "2", plot = FALSE),
  "1"   = smoothingSelection(data_nchi, penalty_type = "1", plot = FALSE),
  "0"   = smoothingSelection(data_nchi, penalty_type = "0", plot = FALSE)
)

# One valid way to obtain the result is to provide the optimal smoothing result:

opt <- optimalSmoothing(res_list)
result <- plotOptimalFits(data = data_nchi, opt_res = opt, graph = TRUE)

# The other valid way is to provide the smoothed object directly:

plot_opt <- plotOptimalSmoothing(data = data_nchi, opt_result = opt)
result <- plotOptimalFits(data = data_nchi, smoothed = plot_opt$smoothed, graph = TRUE)
}

}
