% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectBestPartitions.R
\name{selectBestPartitions}
\alias{selectBestPartitions}
\title{Select Best Partitions Based on Top Ranking Criteria}
\usage{
selectBestPartitions(civf_list, clustering_set, graph = FALSE)
}
\arguments{
\item{civf_list}{A list returned by \code{buildCIVf()}, containing validation matrices and optimization rules.}

\item{clustering_set}{A list returned by \code{runClusteringRange()}, containing repeated clustering results.}

\item{graph}{Logical. If \code{TRUE}, shows a bar chart of top-k frequencies (default: \code{FALSE}).}
}
\value{
A named list of selected partitions, each containing:
\describe{
\item{k}{The number of clusters.}
\item{criterion}{The criterion by which the partition was selected.}
\item{index}{The index of the selected repetition.}
\item{partition}{An integer vector representing the cluster assignment.}
}
The list has class \code{"ccccBestPartitions"}.
}
\description{
This function selects the best clustering partitions that ranked first (Top-1)
in at least one internal validation criterion for each value of \code{k}.
For each selected partition, it returns the optimal repetition that maximizes
(or minimizes) the associated criterion value, depending on the optimization rule.
}
\details{
Optionally, it displays a bar plot showing how frequently each number of clusters
appears among the top-ranked solutions.
}
\examples{
\dontrun{
tdm <- system.file("extdata", "tdm.csv", package = "cccc")
corpus <- system.file("extdata", "corpus.csv", package = "cccc")
data <- importData(tdm_file = tdm, corpus_file = corpus,
sep_tdm = ";",sep_corpus_info = ";",zone="stat")

data_nchi <- normalization(data, normty = "nchi", sc = 1000)
g <- runClusteringRange(data_nchi,
k_range = 2:26,
n_repeats = 20,
seed = 123,
dist_method = "euclidean",
verbose = TRUE)

g2 <- buildCIVf(clustering_set=g,
criteria_set = NULL,
min_valid_values = 0.75,
verbose = TRUE)

g3 <- selectBestPartitions(g2, g, graph=T)
}

}
