% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvePlot.R
\name{curvePlot}
\alias{curvePlot}
\title{Plot Temporal Curves of Keyword Frequencies}
\usage{
curvePlot(
  data,
  r = 1,
  themety = "light",
  size_class = NULL,
  x_leg = 0.85,
  x_lab = "years"
)
}
\arguments{
\item{data}{A list containing the output of \code{importData()}}

\item{r}{Scalar. Interval for thinning x-axis year labels (default = 1).}

\item{themety}{Character. Theme type: "light" (default) or "dark".}

\item{size_class}{Numeric vector for line sizes by class (default is context-dependent).}

\item{x_leg}{Numeric. x-position of legend (default = 0.85).}

\item{x_lab}{Character. Label for x-axis (default = "year").}
}
\description{
This function generates a temporal plot of keyword frequencies (normalized or raw)
from a TDM in long format. It allows visualizing curves grouped by frequency zone,
with thematic customization options.
}
\examples{
\donttest{
tdm <- system.file("extdata", "tdm.csv", package = "cccc")
corpus <- system.file("extdata", "corpus.csv", package = "cccc")
data <- importData(tdm_file = tdm, corpus_file = corpus,
sep_tdm = ";",sep_corpus_info = ";",zone="stat")

curvePlot(data)
}

}
