% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facetPlot.R
\name{facetPlot}
\alias{facetPlot}
\title{Create a ggplot of term curves by faceting}
\usage{
facetPlot(
  data,
  keyword_selection = list(type = "frequency", n = 3, kw.list = NULL),
  r = 4,
  scales = "fixed",
  leg = TRUE,
  themety = "light",
  size_class = NULL,
  x_lab = "year"
)
}
\arguments{
\item{data}{A (molten) dataframe of the term by documents table.}

\item{keyword_selection}{A list...}

\item{r}{A scalar for thinning labels of time x-scale.}

\item{scales}{Type of scale for facets: "fixed" (the default) or "free"}

\item{leg}{Whether legend must be drawn (T is default) or not}

\item{themety}{Type of theme.}

\item{size_class}{size of lines per frequency class.}

\item{x_lab}{x-axis label: "year" (default), user can set a more complete label (e.g., "year/volume").}

\item{norm}{Logical value: true (by default) if frequencies have been normalized.}
}
\value{
A ggplot object.
}
\description{
Draw curves of term frequency.
}
\examples{
# facetPlot(dat_df)
}
