% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facetPlot.R
\name{facetPlot}
\alias{facetPlot}
\title{Faceted Plot of Keyword Frequency Curves}
\usage{
facetPlot(
  data,
  keyword_selection = list(type = "frequency", n = 3, kw.list = NULL),
  r = 4,
  scales = "fixed",
  leg = TRUE,
  themety = "light",
  size_class = NULL,
  x_lab = "year"
)
}
\arguments{
\item{data}{A list object returned by \code{importData()}, containing the term-document matrix in long format and metadata.}

\item{keyword_selection}{A list specifying how to select the keywords to highlight. The list should contain:
\describe{
\item{\code{type}}{ (character) One of \code{"frequency"}, \code{"random"}, or \code{"list"} to choose keywords based on total frequency, at random, or from a user-defined list.}
\item{\code{n}}{ (integer) Number of keywords to select per zone (used for \code{type = "frequency"} or \code{"random"}).}
\item{\code{kw.list}}{ (character vector) List of specific keywords to plot (used for \code{type = "list"}).}
}}

\item{r}{An integer that controls the thinning of x-axis year labels (e.g., every \code{r}-th year is labeled).}

\item{scales}{Character. Whether y-axis scales should be \code{"fixed"} (default) or \code{"free"} across facets.}

\item{leg}{Logical. Whether to display the legend (\code{TRUE} by default).}

\item{themety}{Character. Theme type: \code{"light"} (default) or \code{"dark"}.}

\item{size_class}{Optional numeric vector. Specifies the relative thickness of lines by zone. Defaults are assigned if not provided.}

\item{x_lab}{Character. Label for the x-axis. Defaults to \code{"year"} but can be customized (e.g., \code{"year/volume"}).}
}
\value{
A \code{ggplot} object showing smoothed or raw frequency curves of selected keywords, grouped by zone.
}
\description{
This function creates a faceted \code{ggplot2} visualization of keyword frequency curves across time, grouped by frequency zone. It allows selecting a subset of keywords either by random sampling, by frequency, or by a custom list. The function returns a \code{ggplot} object where each panel corresponds to a frequency zone, and selected keywords are highlighted with distinct colors.
}
\examples{
\dontrun{
tdm <- system.file("extdata", "tdm.csv", package = "cccc")
corpus <- system.file("extdata", "corpus.csv", package = "cccc")
data <- importData(tdm_file = tdm, corpus_file = corpus,
sep_tdm = ";",sep_corpus_info = ";",zone="stat")

facetPlot(data, keyword_selection = list(type="frequency", n=3, kw.list=NULL),
r = 4, scales = "free", leg = TRUE, themety = "light", size_class = NULL, x_lab = "year")
}

}
