% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMassPlot.R
\name{rowMassPlot}
\alias{rowMassPlot}
\title{Plot Row Masses by Frequency Zone}
\usage{
rowMassPlot(data)
}
\arguments{
\item{data}{A list contained the output of \code{importData()}}
}
\value{
A \code{ggplot2} object showing the distribution of total frequencies by zone.
The bars are colored by zone and frequency interval.
}
\description{
This function creates a bar plot of keywords ordered by their total frequency
and colored by their assigned frequency zone, combining both the zone
and its frequency interval label.
}
\examples{
\dontrun{
tdm <- system.file("extdata", "tdm.csv", package = "cccc")
corpus <- system.file("extdata", "corpus.csv", package = "cccc")
data <- importData(tdm_file = tdm, corpus_file = corpus,
sep_tdm = ";",sep_corpus_info = ";",zone="stat")

rowMassPlot(tdm = data$tdm)
}

}
